requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
Flask==2.0.0
gunicorn==20.1.0
idna==2.10
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.0
MarkupSafe==2.0.0
multidict==5.1.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==2.0.1
yarl==1.6.3
